Adding peer dependencies required for juliancwirko:postcss
Creating a new project will result in errors about peer dependencies if following this guide. I took a look at juliancwirko:postcss on github and there was a couple of other things that were added to get things up and running. I just figured it should be reflected in this documentation as well.